% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_snpdata.R
\name{import_snpdata}
\alias{import_snpdata}
\title{Import data for SNP checkk}
\usage{
import_snpdata(input_folder)
}
\arguments{
\item{input_folder}{A folder with VCF files from SNP data}
}
\value{
Dataframe of formatted SNP data
}
\description{
Import data for SNP checkk
}
