% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_snpcheck.R
\name{perform_snpcheck}
\alias{perform_snpcheck}
\title{Perform SNP check}
\usage{
perform_snpcheck(
  snp_table,
  study = NULL,
  sample = NULL,
  vecode = NULL,
  usable_minimum = 18,
  identical_minimum = 18,
  distance_max = 4,
  check_blood = F
)
}
\arguments{
\item{snp_table}{A number}

\item{study}{Study identifier to include}

\item{sample}{Sample identifier to include}

\item{usable_minimum}{Minimum usable SNP's to include in analysis}

\item{identical_minimum}{Minimum identical SNP's to determine similarity}

\item{distance_max}{Maximun distance of changing SNP loci}
}
\value{
Dataframe of pairwise comparisons
}
\description{
Perform SNP check
}
